move IndexUid to dedicated module
